==> scRNA-seq/5-analysis_additional_samples/7.1-transfer_label_validation_cohort_myeloid.Rmd <==
---
title: 'Myeloid: Label and coordinate transfer (validation cohort)'
author: "Ramon Massoni-Badosa"
date: "`r Sys.Date()`"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```


# Introduction

Here, we will transfer the label and UMAP coordinates from the discovery cohort to the validation cohort. We will do it one compartment at a time (e.g. myeloid), so we can validate that the transfer is robust

Specifically, we will follow these steps for every compartment:

1. Include or update the annotation label for this object.
2. Plot the integrated object split by type (quey/reference) to visualize if discovery and validation cohorts were integrated properly.
3. Transfer label.
4. Transfer UMAP coordinates + visualize
5. Exclude cells with a low annotation probability and/or a high doublet score
6. Visualize heatmap of specific markers to validate that the cluster and annotation holds for the new cohort.

# Pre-processing

## Load packages

```{r}
library(Seurat)
library(scCustomize)
library(tidyverse)
library(propr)
library(caret)
library(class)
library(harmony)
library(here)
library(glue)
library(DT)
```


## Source functions

```{r}
source(here("scRNA-seq/bin/SLOcatoR_functions.R"))
source(here("scRNA-seq/bin/utils.R"))
source(here("scRNA-seq/bin/utils_final_clusters.R"))
source(here("scRNA-seq/bin/utils_figure4.R"))
color_palette <-  c("#1CFFCE", "#90AD1C", "#C075A6", "#85660D", "#5A5156", "#AA0DFE",
                    "#F8A19F", "#F7E1A0", "#1C8356", "#FEAF16", "#822E1C", "#C4451C",
                    "#1CBE4F", "#325A9B", "#F6222E", "#FE00FA", "#FBE426", "#16FF32",
                    "black",   "#3283FE", "#B00068", "#DEA0FD", "#B10DA1", "#E4E1E3",
                    "#90AD1C", "#FE00FA", "#85660D", "#3B00FB", "#822E1C", "coral2",
                    "#1CFFCE", "#1CBE4F", "#3283FE", "#FBE426", "#F7E1A0", "#325A9B",
                    "#2ED9FF", "#B5EFB5", "#5A5156", "#DEA0FD", "#FEAF16", "#683B79",
                    "#B10DA1", "#1C7F93", "#F8A19F", "dark orange", "#FEAF16", "#FBE426",
                    "Brown")
```


# Myeloid

```{r}
# Read data, clean and include annotation
myeloid <- readRDS(here("scRNA-seq/results/R_objects/seurat_objects_revision/4.1-myeloid_seurat_object_discovery_validation_cohorts.rds"))
myeloid <- updateAnnotation(myeloid)
myeloid$annotation_20230508 <- myeloid$annotation_20220619
myeloid$annotation_20230508[myeloid$annotation_20230508 == "myeloid_multiome"] <- "unannotated" 
# myeloid$annotation_20230508[myeloid$annotation_20230508 == "myeloid_multiome"] <- "unannotated" 
myeloid$annotation_20230508[is.na(myeloid$annotation_20230508)] <- "unannotated"
myeloid$annotation_20230508 <- factor(
  myeloid$annotation_20230508,
  levels = names(colors_20230508[["myeloid"]])
)
Idents(myeloid) <- "annotation_20230508"
myeloid$is_reference <- ifelse(
  myeloid$cohort_type == "discovery" & myeloid$assay == "3P",
  "reference",
  "query"
)
DimPlot(
  myeloid,
  group.by = "annotation_20230508",
  split.by = "is_reference",
  cols = colors_20230508[["myeloid"]]
)

# Split training and test sets
data_sets <- split_training_and_test_sets(
  myeloid,
  split_var = "is_reference",
  referece_label = "reference",
  query_label = "query",
  reduction = "harmony",
  n_dims = 25
)


# Transfer label
annotation_query_df <- transfer_label(
  seurat_obj = myeloid,
  training_set = data_sets$training_set,
  test_set = data_sets$test_set,
  k = 5,
  response_var = "annotation_20230508"
)

# Transfer UMAP coords
umap_test_df <- transfer_umap_coords(
  seurat_obj = myeloid,
  training_set = data_sets$training_set,
  test_set = data_sets$test_set,
  umap1_var = "UMAP_1_20220215",
  umap2_var = "UMAP_2_20220215",
  k = 15
)


# Include annotation and UMAP coords in Seurat object
myeloid$annotation_20230508[annotation_query_df$query_cells] <- annotation_query_df$annotation
Idents(myeloid) <- "annotation_20230508"
myeloid$annotation_20230508_probability <- NA
myeloid$annotation_20230508_probability[annotation_query_df$query_cells] <- annotation_query_df$annotation_prob
myeloid$UMAP_1_20220215[umap_test_df$query_cells] <- umap_test_df$UMAP1
myeloid$UMAP_2_20220215[umap_test_df$query_cells] <- umap_test_df$UMAP2
p1 <- FeaturePlot(
  myeloid[, annotation_query_df$query_cells],
  "annotation_20230508_probability"
) + scale_color_viridis_c(option = "magma")
p2 <- FeaturePlot(
  myeloid[, annotation_query_df$query_cells],
  "doublet_score_scDblFinder"
) + scale_color_viridis_c(option = "magma")
p1 | p2
DimPlot(
  myeloid,
  group.by = "annotation_20230508",
  split.by = "is_reference",
  cols = colors_20230508[["myeloid"]]
)
table(myeloid$annotation_20230508_probability)
```


We observe 3 or 4 clusters of cells with a high doublet score and low annotation probability. Let us find and exclude them:

```{r}
myeloid <- FindNeighbors(myeloid, reduction = "harmony", dims = 1:25)
myeloid <- FindClusters(myeloid, resolution = 0.1)
DimPlot(myeloid, cols = color_palette)
```

Clusters 3, 9:

```{r}
markers <- map_df(c("3",  "9"), \(x) {
  df <- FindMarkers(myeloid, ident.1 = x, only.pos = TRUE, logfc.threshold = 0.75)
  df$cluster <- x
  df$gene <- rownames(df)
  df
})
DT::datatable(markers)
```


Cluster 3: B-cell doublets
Cluster 9: T-cell doublets

```{r}
cells_to_exclude <- 
  myeloid$cohort_type == "validation" &
  myeloid$seurat_clusters %in% c("3", "9")
myeloid <- myeloid[, !cells_to_exclude]
DimPlot(myeloid, cols = color_palette, split.by = "is_reference")
myeloid <- RunUMAP(myeloid, dims = 1:25, reduction = "harmony")
Idents(myeloid) <- "annotation_20230508"
DimPlot(
  myeloid,
  group.by = "annotation_20230508",
  split.by = "is_reference",
  cols = colors_20230508[["myeloid"]]
)
```


Plot dot plot

```{r}
goi_myeloid <- c(goi_myeloid, "SELENOP", "APOE", "MMP9", "MMP12", "C1QA", "ITGAX", "ZEB2")
dot_plot_validation <- DotPlot(myeloid[, myeloid$cohort_type == "validation"], features = rev(goi_myeloid)) +
  # coord_flip() +
  scale_color_distiller(palette = "Blues", direction = 1) +
  coord_flip() +
  theme(
    axis.title = element_blank(),
    axis.text.y = element_text(size = 6),
    legend.text = element_text(size = 5.5),
    legend.title = element_text(size = 6),
    axis.text.x = element_text(size = 6, angle = 45, hjust = 1, vjust = 1))
dot_plot_validation


# Violin plot
p <- Stacked_VlnPlot(
  seurat_object = myeloid,
  features = goi_myeloid,
  x_lab_rotate = TRUE,spacing_unit = "mm",
  plot_legend = TRUE,
  colors_use = c("#004080", "#006600"),
  # colors_use = colors_20230508$myeloid,
  split.by = "cohort_type"
)
p &
  theme(
    axis.title = element_text(size = 7),
    axis.text = element_text(size = 6),
    legend.text = element_text(size = 7))
```


# Save

```{r}
saveRDS(myeloid, here("scRNA-seq/results/R_objects/seurat_objects_revision/5.1-myeloid_seurat_object_discovery_validation_cohorts_annotation.rds"))
```


# Session Information

```{r}
sessionInfo()
```
